remove: in-crate derive module
relocate macros to bn256 and macros.rs
