--- curves/src/derive/field/mod.rs
@@ -1,4 +1,0 @@
-#[macro_use]
-pub mod common;
-#[macro_use]
-pub mod tower;
